requirements.txt
psycopg2-binary
pandas
sql-metadata
pglast